syntax

+++ src/api/v1/utils/ContentService.php
@@ -88,7 +88,11 @@ class ContentService
             $destdir = $this->databasedir.'/'.$type.'/archives';
             $destfile = $destdir . '/' . $keyvalue.'.json';
 
-                
+            // create backup ?
+
+            if (file_exists($file)) {
+                rename($file, $destfile);
+
                 unlink($file);
 
                 $response->setCode(200);
